Update egll.py
Updates to support Venus OS v3.52 changes 
Updates to support dbus-serialbattery v1.5.20241215

Updated to report all BMS's in chain to Venus OS as single battery. Enable tracking of BMS values to all be displayed independently and not as a rolled up sum. 

Requires MODBUS_ADDRESSES to be defined with the Hex ID of the units in the chain. Values should be in Hex